Add files via upload
This file contains my analysis for qPCR done on three samples with 8 target primers. To normalize the data, I use the Ct mean of a sample's control target and subtract the Ct mean of the other targets of interest. 2 is raised to the power of this value, delta, to get the normalized value of RNA expression to the control. This is automatically repeated for all controls defined in the list controls.